+++ tclpkg/tclstubs/tclInt.h
@@ -1,0 +1,19 @@
+/*************************************************************************
+ * Copyright (c) 2011 AT&T Intellectual Property 
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v2.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v20.html
+ *
+ * Contributors: Details at https://graphviz.org
+ *************************************************************************/
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/* dummy header in case configure can't find a real one */
+
+#ifdef __cplusplus
+}
+#endif